requirements.txt
Flask
PyYAML==5.4.1